code rearrangement
