+++ Event.h
@@ -6,10 +6,10 @@ struct Event // represents a button press/release
     enum type
     {
         Press,
-        Release,
+        Release
     };
     type eventType;
     uint8_t code;
-    unsigned long wait = 0; 
+    uint16_t wait = 0; 
     bool Handled = false;
 };
