+++ MacroProcessor.h
@@ -4,28 +4,32 @@
 #include "Event.h"
 #include "KeyCode.h"
 
+#include "EventProcessor.h"
 
-class MacroProcessor
+
+class MacroProcessor : public EventProcessor
 {
-  
+  private:  
+    enum State
+    {
+      IDLE = 0,
+      RECORDING,
+    };
   public:
     MacroProcessor();
     virtual ~MacroProcessor();  
-    void DispatchEvent(Event& e);
+    void DispatchEvent(Event& e) override;
 
   private:
     void executeMacro(uint8_t macro);
     void recordEvent(Event& e);
     void startRecording();
     
-    enum State
-    {
-      IDLE = 0,
-      RECORDING,
-    };  
+  private:
+    static constexpr uint8_t MAX_SIZE = 128;
     QueueArray<Event>* m_EventQueue;
     State m_State = IDLE;
-    unsigned long m_LastEventTime; 
+    unsigned long int m_LastEventTime; 
 };
 
 #endif
