+++ PlatformAPI.h
@@ -1,9 +1,12 @@
 #pragma once
+
 #include <stdint.h>
 #include "InputHandler.h"
 #include "Event.h"
 
-enum class PlatformAPIType
+#include "EventProcessor.h"
+
+enum class PlatformAPIType 
 {
     None,
     XInput,
@@ -11,13 +14,14 @@ enum class PlatformAPIType
     Dualshock
 };
 
-class PlatformAPI {
+class PlatformAPI : public EventProcessor 
+{
     public:
         virtual void Press(uint8_t code) = 0;
         virtual void Release(uint8_t code) = 0;
         virtual void SetJoyStick(JoyStickType type, int16_t x, int16_t y) = 0 ;
 
-        void DispatchEvent(Event& e);
+        void DispatchEvent(Event& e)override;
         
     public:
         static PlatformAPIType Current() {return s_CurrentAPI;}
